modularized planner canvas
